refactor: change naming to one style
