+++ src/features/courses/single-course/components/skeletons/table.tsx
