+++ src/features/courses/single-course/components/skeletons/typography.tsx
