+++ src/features/courses/single-course/topic/components/content-panel.tsx
